@@ -1,50 +1,49 @@
 using Microsoft.Extensions.DependencyInjection;
 
-namespace Crip.AspNetCore.Logging.Tests
+namespace Crip.AspNetCore.Logging.Tests.Services;
+
+public class TimeMeasurableTests
 {
-    public class TimeMeasurableTests
+    private readonly Mock<IStopwatch> _stopwatch = new();
+    private readonly IServiceProvider _provider;
+    private readonly IMeasurable _measurable;
+
+    public TimeMeasurableTests()
     {
-        private readonly Mock<IStopwatch> _stopwatch = new();
-        private readonly IServiceProvider _provider;
-        private readonly IMeasurable _measurable;
+        _provider = new ServiceCollection()
+            .AddSingleton(_stopwatch.Object)
+            .BuildServiceProvider();
 
-        public TimeMeasurableTests()
-        {
-            _provider = new ServiceCollection()
-                .AddSingleton(_stopwatch.Object)
-                .BuildServiceProvider();
+        _measurable = new TimeMeasurable(_provider);
+    }
 
-            _measurable = new TimeMeasurable(_provider);
-        }
+    [Fact, Trait("Category", "Unit")]
+    public void StartMeasure_CreatesNewInstance()
+    {
+        var result = _measurable.StartMeasure();
 
-        [Fact, Trait("Category", "Unit")]
-        public void StartMeasure_CreatesNewInstance()
-        {
-            var result = _measurable.StartMeasure();
+        result
+            .Should().NotBeNull()
+            .And.NotBe(_measurable);
+    }
 
-            result
-                .Should().NotBeNull()
-                .And.NotBe(_measurable);
-        }
+    [Fact, Trait("Category", "Unit")]
+    public void StopMeasure_ThrowsErrorIfNotStarted()
+    {
+        Action act = () => _measurable.StopMeasure();
 
-        [Fact, Trait("Category", "Unit")]
-        public void StopMeasure_ThrowsErrorIfNotStarted()
-        {
-            Action act = () => _measurable.StopMeasure();
+        act.Should()
+            .Throw<ArgumentNullException>()
+            .WithMessage("Could not stop not started time measurement. (Parameter '_stopwatch')");
+    }
 
-            act.Should()
-                .Throw<ArgumentNullException>()
-                .WithMessage("Could not stop not started time measurement. (Parameter '_stopwatch')");
-        }
+    [Fact, Trait("Category", "Unit")]
+    public void StopMeasure_StopsStopwatch()
+    {
+        var measure = _measurable.StartMeasure();
+        var result = measure.StopMeasure();
 
-        [Fact, Trait("Category", "Unit")]
-        public void StopMeasure_StopsStopwatch()
-        {
-            var measure = _measurable.StartMeasure();
-            var result = measure.StopMeasure();
-
-            _stopwatch.Verify(stopwatch => stopwatch.Stop(), Times.Once());
-            result.Should().Be(_stopwatch.Object);
-        }
+        _stopwatch.Verify(stopwatch => stopwatch.Stop(), Times.Once());
+        result.Should().Be(_stopwatch.Object);
     }
 }
