@@ -6,7 +6,13 @@ builder.Host.UseSerilog((context, configuration) => configuration
     .WriteTo.Console(outputTemplate: "{Timestamp:o} [{Level:u3}] {Message}{NewLine}{Properties}{NewLine}{Exception}")
     .ReadFrom.Configuration(context.Configuration));
 
-builder.Services.AddRequestLogging();
+builder.Services.AddRequestLogging(options =>
+{
+    options.LogResponse = true;
+    options.AuthorizationHeaders.AuthorizationHeaderNames.Add("x-authorize");
+    options.LongJsonContent.MaxCharCountInField = 255;
+    options.LongJsonContent.LeaveOnTrimCharCountInField = 15;
+});
 
 builder.Services.AddControllers();
 // Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
